requirements.txt
streamlit
snowflake-connector-python
snowflake-connector-python[pandas]
openai
pandas
numpy
google
protobuf